+++ tests/qft.rs
@@ -61,7 +61,7 @@ fn qft(input_state: ProductState) -> Option<SuperPosition> {
     }
 
     mini_circuit
-        .change_register(input_state.into_super_position())
+        .change_register(input_state.into())
         .unwrap()
         .simulate();
